backend/requirements.txt
django>=6.0.1
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.5.1
psycopg2-binary>=2.9.11
Pillow>=10.0.0
django-cors-headers>=4.0.0
openpyxl>=3.1.2

